requirements.txt
torch
transformers
python-dotenv
SentencePiece
streamlit 
groq

fastapi
uvicorn

